fix: fix date timezone

+++ src/utils/date.ts
@@ -1,7 +1,7 @@
 export const formatLongDate = (date: Date) =>
   new Intl.DateTimeFormat('en-GB', {
     dateStyle: 'full',
-    timeStyle: 'short'
+    timeStyle: 'long'
   }).format(date)
 
 export const formatShortDate = (date: Date) =>
